--- lib/presentation/home/language_selection/binding/language_selection_binding.dart
@@ -1,10 +1,0 @@
-import 'package:get/get.dart';
-
-import '../controller/language_selection_controller.dart';
-
-class LanguageSelectionBinding extends Bindings {
-  @override
-  void dependencies() {
-    Get.lazyPut(() => LanguageSelectionController());
-  }
-}
