+++ lib/presentation/source_target_language/binding/source_target_langugage_binding.dart
@@ -1,0 +1,10 @@
+import 'package:get/get.dart';
+
+import '../controller/source_target_langugage_controller.dart';
+
+class SourceTargetLanguageBinding extends Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut(() => SourceTargetLanguageController());
+  }
+}
